+++ swagger_parser/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.39.0
 - Handle request body name generation case when request body is declared within request
+- Fix incorrect deserialization syntax for undiscriminated unions (Freezed serializer)
 
 ## 1.38.0
 - Add `replacement_rules_for_raw_schema` option for raw schema objects replacement rules